python/requirements.txt
numpy
opencv-python
rich
torch
ultralytics
yacs
termcolor
tabulate
tensorboard
fvcore
gdown
scikit-learn
fastreid
faiss-cpu
boxmot
xformers
